requirements.txt
# Web Scraping
beautifulsoup4>=4.9.3
requests>=2.25.1
google>=3.0.0

# Data Processing
pandas>=1.2.4
tiktoken>=0.8.0

# LLM & AI
openai>=1.61.1
langchain-openai>=0.0.2

# Development
python-dotenv>=0.19.0
jupyter>=1.0.0

werkzeug==2.0.1
google-api-python-client

# FastAPI
fastapi>=0.100.0,<1.0.0
pydantic>=2.0.0,<3.0.0
uvicorn>=0.15.0,<1.0.0
aiohttp
        
fake-useragent==2.0.3
langchain==0.3.17